Bcc.Node.Protocol.Cole

Documentation

mkSomeConsensusProtocolCole ∷ NodeColeProtocolConfiguration → Maybe ProtocolFilepaths → ExceptT ColeProtocolInstantiationError IO SomeConsensusProtocol

Make SomeConsensusProtocol using the Cole instance.

This lets us handle multiple protocols in a generic way.

This also serves a purpose as a sanity check that we have all the necessary type class instances available.

Errors

data ColeProtocolInstantiationError

Constructors

CanonicalDecodeFailure !FilePath !Text

GenesisHashMismatch !GenesisHash !GenesisHash

DelegationCertificateFilepathNotSpecified

GenesisConfigurationError !FilePath !ConfigurationError

GenesisReadError !FilePath !GenesisDataError

CredentialsError !ColeLeaderCredentialsError

SigningKeyDeserialiseFailure !FilePath

SigningKeyFilepathNotSpecified

Reusable parts

readGenesis ∷ GenesisFile → Maybe GenesisHash → RequiresNetworkMagic → ExceptT ColeProtocolInstantiationError IO Config

readLeaderCredentials ∷ Config → Maybe ProtocolFilepaths → ExceptT ColeProtocolInstantiationError IO (Maybe ColeLeaderCredentials)