Bcc.Tracing.Metrics

Documentation

data KESMetricsData

KES-related data to be traced as metrics.

Constructors

NoKESMetricsData
The current protocol does not support KES.

TOptimumKESMetricsData

Fields

!Period

The current KES period of the hot key, relative to the start KES period of the operational certificate.

!MaxKESEvolutions

The configured max KES evolutions.

!OperationalCertStartKESPeriod

The start KES period of the configured operational certificate.

newtype MaxKESEvolutions

The maximum number of evolutions that a KES key can undergo before it is considered expired.

Constructors

MaxKESEvolutions Word64

newtype OperationalCertStartKESPeriod

The start KES period of the configured operational certificate.

Constructors

OperationalCertStartKESPeriod Period

class HasKESMetricsData blk where

Minimal complete definition

Nothing

Methods

getKESMetricsData ∷ Proxy blk → ForgeStateInfo blk → KESMetricsData

class HasKESInfo blk where

Minimal complete definition

Nothing

Methods

getKESInfo ∷ Proxy blk → ForgeStateUpdateError blk → Maybe KESInfo

data ForgingStats

This structure stores counters of blockchain-related events, per individual forge thread. These counters are driven by traces.

Constructors

ForgingStats

Fields

fsTxsProcessedNum ∷ !(IORef Int)

Transactions removed from mempool.

fsState ∷ !(TVar (Map ThreadId (TVar ForgeThreadStats)))

fsBlocksUncoupled ∷ !(TVar Int64)

Blocks forged since last restart not on the current chain

data ForgeThreadStats

Per-forging-thread statistics.

Constructors

ForgeThreadStats

Fields

ftsNodeCannotForgeNum ∷ !Int

ftsNodeIsLeaderNum ∷ !Int

ftsBlocksForgedNum ∷ !Int

ftsSlotsMissedNum ∷ !Int

Potentially missed slots. Note that this is not the same as the number of missed blocks, since this includes all occurences of not reaching a leadership check decision, whether or not leadership was possible or not.

Also note that when the aggregate total for this metric is reported in the multi-pool case, it can be much larger than the actual number of slots occuring since node start, for it is a sum total for all threads.

ftsLastSlot ∷ !Int

mapForgingCurrentThreadStats ∷ ForgingStats → (ForgeThreadStats → (ForgeThreadStats, a)) → IO a

mapForgingCurrentThreadStats_ ∷ ForgingStats → (ForgeThreadStats → ForgeThreadStats) → IO ()

mapForgingStatsTxsProcessed ∷ ForgingStats → (Int → Int) → IO Int

mkForgingStats ∷ IO ForgingStats

threadStatsProjection ∷ ForgingStats → (ForgeThreadStats → a) → IO [a]